Support high-availability replicas
Use redis redlock to compete for a lock for each message ID. The bot
that manages to take the lock will respond to that message.

Currently this breaks the startwatch functionality since it depends on
state in a specific bot instance. That state will need to be migrated to
external storage for this to work properly.